refactor(transform): extract flag-driven thinking compact out of smart_compact
The SmartCompact scenario flag's actual behavior is a thinking trim —
removing thinking blocks from historical conversation rounds — not a
general conversation compression. Its implementation lived in
internal/smart_compact, a package otherwise dedicated to the vmodel
compact models' compression strategies, which made the flag path
confusingly reach into vmodel territory.

- Move the transform to internal/server/transform.ThinkingCompactTransform
  (server-domain layer, alongside the other flag/scenario-driven transforms
  like clean_header and max_tokens). Name() stays "compact_thinking" so
  recorded TransformSteps and probe output are unchanged.
- The flag path (protocol_transform.go) and the "compact-thinking" virtual
  model (vmodel/anthropic/defaults.go) now build the same transform from its
  new home; internal/server no longer imports smart_compact at all.
- Delete from smart_compact: thinking_transform.go, the legacy duplicate
  CompactTransformer (compact.go, test-only per its own doc), and
  compact_helpers.go (helpers were exclusive to the thinking trim; the dead
  CompressionStrategy interface goes with them). Port the thinking tests to
  internal/server/transform, rewritten against the Transform API entry point.
- Add smart_compact/doc.go scoping the package to the vmodel compression
  strategies, and point .design/rule-flags.md's smart_compact row at the new
  implementation home.

go build ./... clean; server/transform, smart_compact, vmodel, and
internal/server test suites pass; harness matrix 327/327 and replay batch
8/8 (1 pre-existing skip) stay green.
